pyproject.toml
[project]
name = "rag-pipeline"
version = "1.0.0"
description = "Production-grade RAG Pipeline using Free Cloud Services"
requires-python = ">=3.10"

dependencies = [
    # ============================================
    # API FRAMEWORK
    # ============================================
    # FastAPI: Modern web framework for building APIs
    # - Automatic OpenAPI documentation
    # - Type validation with Pydantic
    # - Async support
    "fastapi>=0.109.0",
    
    # Uvicorn: ASGI server to run FastAPI
    # - High performance
    # - Hot reload in development
    "uvicorn[standard]>=0.27.0",
    
    # For handling file uploads in FastAPI
    "python-multipart>=0.0.6",
    
    # ============================================
    # VECTOR DATABASE
    # ============================================
    # PyMilvus: Python SDK for Milvus/Zilliz
    # - Connects to your Zilliz Cloud cluster
    # - Handles vector operations
    "pymilvus>=2.3.0",
    
    # ============================================
    # CLOUD API CLIENTS
    # ============================================
    # HTTPX: Modern HTTP client
    # - Used for HuggingFace API calls
    # - Async support
    "httpx>=0.26.0",
    
    # Groq: Official Groq client
    # - Fast LLM inference
    # - OpenAI-compatible API
    "groq>=0.4.0",
    
    # Cohere: Official Cohere client
    # - Reranking API
    # - Also has embeddings (backup option)
    "cohere>=4.45.0",
    
    # ============================================
    # DOCUMENT PROCESSING
    # ============================================
    # PyPDF: Extract text from PDF files
    "pypdf>=3.17.0",
    
    # python-docx: Extract text from Word documents
    "python-docx>=1.1.0",
    
    # BeautifulSoup: Parse HTML content
    "beautifulsoup4>=4.12.0",
    
    # Aiofiles: Async file operations
    "aiofiles>=23.2.0",
    
    # ============================================
    # TEXT PROCESSING
    # ============================================
    # NLTK: Natural Language Toolkit
    # - Sentence tokenization
    # - Text processing utilities
    "nltk>=3.8.0",
    
    # ============================================
    # CONFIGURATION
    # ============================================
    # Pydantic: Data validation
    # - Type checking
    # - Serialization
    "pydantic>=2.5.0",
    
    # Pydantic Settings: Load config from environment
    "pydantic-settings>=2.1.0",
    
    # Python-dotenv: Load .env files
    "python-dotenv>=1.0.0",
    
    # ============================================
    # LOGGING
    # ============================================
    # Structlog: Structured logging
    # - JSON output for production
    # - Pretty output for development
    "structlog>=24.1.0",
    
    # ============================================
    # UTILITIES
    # ============================================
    # Tenacity: Retry with exponential backoff
    # - Handle transient API failures
    "tenacity>=8.2.0",
    
    # NumPy: Numerical operations
    # - Vector calculations
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]